requirements.txt
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.4
certifi==2020.4.5.1
chardet==3.0.4
cloudinary==1.20.0
Django==3.0.6
django-cloudinary-storage==0.2.3
django-cors-headers==3.3.0
django-filebrowser-no-grappelli==3.8.0
django-grappelli==2.14.2
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
isort==4.3.21
jsmin==2.2.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2-binary==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.3
python-decouple==3.3
python-magic==0.4.18
pytz==2020.1
requests==2.23.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
